% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\name{merge.rollcall}
\alias{merge.rollcall}
\title{Merge rollcall objects}
\usage{
\method{merge}{rollcall}(x = NULL, y = NULL, rc_list = NULL, by = NULL)
}
\arguments{
\item{x}{An object of class \code{dwnominate()}.}

\item{y}{An object of class \code{dwnominate()}.}

\item{rc_list}{An list of object of \code{dwnominate()} objects.}
}
\description{
Merge rollcall objects
}
\examples{
library(pscl) # for the rollcall summary method
data(senate)
combined_rcs = merge(senate[[1]], senate[[2]], by='icpsrLegis')
summary(combined_rcs)

all_rcs = merge.rollcall(rc_list=senate, by='icpsrLegis')
summary(all_rcs)
}
