% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\name{merge_rollcalls}
\alias{merge_rollcalls}
\title{Merge rollcall objects}
\usage{
merge_rollcalls(rollcalls, by = NULL)
}
\arguments{
\item{rollcalls}{A list of \code{rollcall} objects.}

\item{by}{The column in legis.data to use as an ID for matching
legislators. If not provided legislators are matched based on
name.}
}
\description{
Merge a list of \code{rollcall} objects.
}
\examples{
data(nhsenate)
combined_rcs <- merge_rollcalls(nhsenate, by='name')
print(combined_rcs)
}
