% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhsen}
\alias{nhsen}
\title{New Hampshire Senate Roll Calls}
\format{
A list of \code{\link[pscl]{rollcall}} objects from the pscl package
}
\source{
\href{http://gencourt.state.nh.us}{New Hampshire General Court}
}
\usage{
nhsen
}
\description{
New Hampshire Senate roll call votes from 1999 to 2019.
}
\examples{
library(pscl)
data(nhsen)
summary(nhsen[[1]])
}
\keyword{datasets}
