% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\name{plot.dwnominate}
\alias{plot.dwnominate}
\title{Plot party means over time}
\usage{
\method{plot}{dwnominate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{dwnominate()}.}

\item{...}{Arguments passed to \code{plot.default()}}
}
\description{
Plot party means over time
}
\examples{
\donttest{
data(nhsen)
results <- dwnominate(nhsen)
plot(results)
}
}
