% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\docType{data}
\name{senate}
\alias{senate}
\title{Sessions 109-113 of the US Senate}
\format{A list of rollcall objects from the pscl package}
\source{
\href{http://www.voteview.com/dwnl.html}{voteview.com}
}
\usage{
senate
}
\description{
Rollcall votes of Senators in the 109-113th sessions of the US
Senate.
}
\examples{
# US Senate data from voteview.com
data(senate)

library(wnominate)
wnom <- wnominate(senate[[1]], polarity = 1:2)
plot(wnom)

\donttest{
dwnom <- dwnominate(senate)
plot(dwnom)
}
}
\keyword{datasets}
