% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_space.R
\name{common_space}
\alias{common_space}
\title{Calculate common space scores}
\usage{
common_space(nom_list, id = NULL, dims = 2, minscale = 5, polarity = NULL)
}
\arguments{
\item{nom_list}{A list of \code{nomObject} results from the
\code{wnominate} package. The number of estimated dimensions in
each \code{nomObject} should be greater than or equal to
\code{dims}.}

\item{id}{Column name in the \code{nomObject}s' \code{legislators}
data frames providing a unique legislator ID. If not specified
row names will be used.}

\item{dims}{The number of dimensions to estimate.}

\item{minscale}{The minimum number of W-NOMINATE scores a
legislator must have to be used for the latent dimension
estimates. Common space scores for legislators with fewer than
\code{minscale} scores will be derived from the estimated
dimensions.}

\item{polarity}{A vector of ID's of length \code{dims} of
legislators who should have relatively positive scores on the
corresponding dimension. If only one ID is provided, it will be
recycled for all dimensions.}
}
\value{
A list of class \code{common space} containing: \itemize{
  \item{legislators} {A data frame of legislator coordinates}
  \item{blackbox} {A list of blackbox results} \item{nom_list} {The
  provided list of \code{wnominate} results} }
}
\description{
The common space function extracts latent dimensions from a
collection of roll call scaling results using the procedure
outlined in Poole (1998) and Poole (2005).
}
\references{
Keith T. Poole. 1998. 'Recovering a Basic Space From a
  Set of Issue Scales.' New York: Cambridge University
  Press. American Journal of Political Science, Vol. 42, No. 3
  (Jul., 1998), pp. 954-993

Keith T. Poole. 2005. 'Spatial Models of Parliamentary Voting.' New
  York: Cambridge University Press.
}
