% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\name{dwnominate}
\alias{dwnominate}
\title{Run DW-NOMINATE}
\usage{
dwnominate(rc_list, id = NULL, wnom = NULL, sessions = NULL, dims = 2,
  model = 1, niter = 4, beta = 5.9539, w = 0.3463, polarity = NULL,
  ...)
}
\arguments{
\item{rc_list}{A list of \code{rollcall} objects from the
\code{pscl} package, in chronological order.}

\item{id}{Column name in the rollcall objects' \code{legis.data}
data frames providing a unique legislator ID. If not specified
legislator names will be used.}

\item{wnom}{A \code{wnominate} or \code{oc} result object (class
\code{nomObject} or \code{OCobject}) providing starting estimates
of legislator ideologies. If not provided, dwnominate will run
\code{wnominate} or \code{oc} to get starting values.}

\item{sessions}{A vector of length 2 providing the first and last
sessions to include. Defaults to \code{c(1, length(rc_list))}.}

\item{dims}{The number of dimensions to estimate. Can be either 1
or 2.}

\item{model}{The degree of the polynomial representing changes in
legislator ideology over time. \code{0} is constant, \code{1} is
linear, \code{2} is quadratic and \code{3} is cubic.}

\item{niter}{Number of iterations. 4 iterations are typically
enough for the results to converge.}

\item{beta}{Starting estimate of the parameter representing the
spatial error in legislator choices.}

\item{w}{Starting estimate for the weight of the second
dimension. The first dimension has a weight of 1, so w should be
<= 1.}

\item{polarity}{A vector of length 1 or \code{dims} specifying, for
each dimension, a legislator who should have a positive
coordinate value. Legislators can be specified either by name or
ID. If unspecified the first legislator in the data is used.}

\item{...}{Arguments passed to \code{wnominate} (if dims==1) or
\code{oc} (if dims>1).}
}
\value{
A list of class \code{dwnominate} containing: \itemize{
  \item{legislators} {A data frame of legislator information}
  \item{rollcalls} {A data frame of rollcall information}
  \item{wnom} {The \code{wnominate} or \code{oc} results used as
  starting points for DW-NOMINATE} \item{rc_list} {The list of
  rollcall objects} }
}
\description{
Run DW-NOMINATE
}
\examples{
# US Senate data originally from voteview.com
data(senate)
results <- dwnominate(senate)
plot(results)
}
\references{
Keith Poole. 2005. 'Spatial Models of Parliamentary
  Voting.' New York: Cambridge University Press.
}
