% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwnominate.R
\name{merge.rollcall}
\alias{merge.rollcall}
\title{Merge rollcall objects}
\usage{
\method{merge}{rollcall}(x = NULL, y = NULL, rc_list = NULL, by = NULL)
}
\arguments{
\item{x}{An object of class \code{dwnominate()}.}

\item{y}{An object of class \code{dwnominate()}.}

\item{rc_list}{An list of object of \code{dwnominate()} objects.}

\item{by}{The column in legis.data to use as an ID for matching
legislators. If not provided legislators are matched based on
name.}
}
\description{
Merge x and y rollcall objects, or a list of rollcall objects.
}
\examples{
data(nhsenate)
combined_rcs = merge(nhsenate[[1]], nhsenate[[2]], by='name')
summary(combined_rcs)
}
